+++ test/unit/main.cpp
@@ -30,6 +30,7 @@ void test_pane();
 void test_textbuf();
 void test_tokenize();
 void test_parse();
+void test_opt();
 void test_path();
 void test_hostfs();
 void test_jsref();
@@ -89,6 +90,7 @@ BRAAM_EXPORT("run_tests") u32 run_tests()
     test_vfs();
     test_tokenize();
     test_parse();
+    test_opt();
     test_procfs();
 
     u32 failures = test_failures();